# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"JsonValue": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"type": "number",
						"format": "double"
					},
					{
						"type": "boolean"
					},
					{
						"$ref": "#/components/schemas/JsonObject"
					},
					{
						"$ref": "#/components/schemas/JsonArray"
					}
				],
				"nullable": true,
				"description": "From https://github.com/sindresorhus/type-fest/\r\nMatches any valid JSON value."
			},
			"JsonObject": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/JsonValue"
				},
				"type": "object",
				"description": "From https://github.com/sindresorhus/type-fest/\r\nMatches a JSON object.\r\nThis type can be useful to enforce some input to be JSON-compatible or as a super-type to be extended from."
			},
			"JsonArray": {
				"description": "From https://github.com/sindresorhus/type-fest/\r\nMatches a JSON array.",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.TipoUsuario": {
				"type": "string",
				"enum": [
					"INDIVIDUAL",
					"EMPRESARIAL"
				]
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"legacyJwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "producttrack",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@mariaximena0724 y emilia"
		}
	},
	"paths": {
...
		"/preferencias-notificaciones/{idUsuario}": {
			"get": {
				"operationId": "ObtenerPreferencias",
				"responses": {
					"200": {
						"description": "Objeto con las preferencias de notificación",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"actualizacion": {
											"type": "boolean"
										},
										"reposicion": {
											"type": "boolean"
										},
										"comentarios": {
											"type": "boolean"
										},
										"productoVencido": {
											"type": "boolean"
										},
										"stockBajo": {
											"type": "boolean"
										}
									},
									"required": [
										"actualizacion",
										"reposicion",
										"comentarios",
										"productoVencido",
										"stockBajo"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Obtiene las preferencias de notificación de un usuario específico.\r\nSi no existen en la base de datos, se devuelven valores por defecto.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/preferencias.controller)",
				"tags": [
					"Preferencias de Notificaciones"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID del usuario",
						"in": "path",
						"name": "idUsuario",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "ActualizarPreferencias",
				"responses": {
					"200": {
						"description": "Mensaje de confirmación",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensaje": {
											"type": "string"
										}
									},
									"required": [
										"mensaje"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Crea o actualiza las preferencias de notificación de un usuario.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/preferencias.controller)",
				"tags": [
					"Preferencias de Notificaciones"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID del usuario",
						"in": "path",
						"name": "idUsuario",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Objeto con las preferencias a modificar",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"actualizacion": {
										"type": "boolean"
									},
									"reposicion": {
										"type": "boolean"
									},
									"comentarios": {
										"type": "boolean"
									},
									"productoVencido": {
										"type": "boolean"
									},
									"stockBajo": {
										"type": "boolean"
									}
								},
								"type": "object",
								"description": "Objeto con las preferencias a modificar"
							}
						}
					}
				}
			}
		},
		"/nutriscan": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Registro creado",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"isTest": {
													"type": "boolean"
												},
												"actualizadoEn": {
													"type": "string",
													"format": "date-time"
												},
												"fechaAnalisis": {
													"type": "string",
													"format": "date-time"
												},
												"tipoAnalisis": {
													"type": "string"
												},
												"respuesta": {
													"$ref": "#/components/schemas/JsonValue"
												},
												"consulta": {
													"type": "string"
												},
												"esAlimento": {
													"type": "boolean"
												},
												"usuarioId": {
													"type": "number",
													"format": "double"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"isTest",
												"actualizadoEn",
												"fechaAnalisis",
												"tipoAnalisis",
												"respuesta",
												"consulta",
												"esAlimento",
												"usuarioId",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Datos inválidos"
					},
					"403": {
						"description": "Acceso denegado"
					}
				},
				"description": "Crea un nuevo análisis de producto usando NutriScan.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/nutriscan.controller)",
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Datos del análisis (sin idUsuario).",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Datos del análisis (sin idUsuario)."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"properties": {
													"isTest": {
														"type": "boolean"
													},
													"actualizadoEn": {
														"type": "string",
														"format": "date-time"
													},
													"fechaAnalisis": {
														"type": "string",
														"format": "date-time"
													},
													"tipoAnalisis": {
														"type": "string"
													},
													"respuesta": {
														"$ref": "#/components/schemas/JsonValue"
													},
													"consulta": {
														"type": "string"
													},
													"esAlimento": {
														"type": "boolean"
													},
													"usuarioId": {
														"type": "number",
														"format": "double"
													},
													"id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"isTest",
													"actualizadoEn",
													"fechaAnalisis",
													"tipoAnalisis",
													"respuesta",
													"consulta",
													"esAlimento",
													"usuarioId",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "Obtiene todos los análisis (admin o desarrollador).\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/nutriscan.controller)",
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/nutriscan/usuario/{usuarioId}": {
			"get": {
				"operationId": "FindByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"allOf": [
													{
														"properties": {
															"usuario": {
																"properties": {
																	"tipoUsuario": {
																		"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
																	},
																	"nombreCompleto": {
																		"type": "string"
																	}
																},
																"required": [
																	"tipoUsuario",
																	"nombreCompleto"
																],
																"type": "object"
															}
														},
														"required": [
															"usuario"
														],
														"type": "object"
													},
													{
														"properties": {
															"isTest": {
																"type": "boolean"
															},
															"actualizadoEn": {
																"type": "string",
																"format": "date-time"
															},
															"fechaAnalisis": {
																"type": "string",
																"format": "date-time"
															},
															"tipoAnalisis": {
																"type": "string"
															},
															"respuesta": {
																"$ref": "#/components/schemas/JsonValue"
															},
															"consulta": {
																"type": "string"
															},
															"esAlimento": {
																"type": "boolean"
															},
															"usuarioId": {
																"type": "number",
																"format": "double"
															},
															"id": {
																"type": "number",
																"format": "double"
															}
														},
														"required": [
															"isTest",
															"actualizadoEn",
															"fechaAnalisis",
															"tipoAnalisis",
															"respuesta",
															"consulta",
															"esAlimento",
															"usuarioId",
															"id"
														],
														"type": "object"
													}
												]
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "Obtiene todos los análisis de un usuario (solo ADMIN).\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/nutriscan.controller)",
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID del usuario.",
						"in": "path",
						"name": "usuarioId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/nutriscan/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"isTest": {
													"type": "boolean"
												},
												"actualizadoEn": {
													"type": "string",
													"format": "date-time"
												},
												"fechaAnalisis": {
													"type": "string",
													"format": "date-time"
												},
												"tipoAnalisis": {
													"type": "string"
												},
												"respuesta": {
													"$ref": "#/components/schemas/JsonValue"
												},
												"consulta": {
													"type": "string"
												},
												"esAlimento": {
													"type": "boolean"
												},
												"usuarioId": {
													"type": "number",
													"format": "double"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"isTest",
												"actualizadoEn",
												"fechaAnalisis",
												"tipoAnalisis",
												"respuesta",
												"consulta",
												"esAlimento",
												"usuarioId",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Datos inválidos"
					}
				},
				"description": "Obtiene todos los análisis de un usuario (solo ADMIN).\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/nutriscan.controller)",
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Elimina un análisis de NutriScan.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/nutriscan.controller)",
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID del análisis.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/notificaciones/preferencias/{idUsuario}": {
			"patch": {
				"operationId": "ActualizarPreferencias",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensaje": {
											"type": "string"
										}
									},
									"required": [
										"mensaje"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Actualiza las preferencias de notificación de un usuario.\r\n\r\n👉 [Documentación general en GitHub](https://mari0724.github.io/producttrack/docs/backend/controllers/notificaciones.controller)",
				"tags": [
					"Notificaciones"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID del usuario.",
						"in": "path",
						"name": "idUsuario",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"actualizacion": {
										"type": "boolean"
									},
									"reposicion": {
										"type": "boolean"
									},
									"comentarios": {
										"type": "boolean"
									},
									"productoVencido": {
										"type": "boolean"
									},
									"stockBajo": {
										"type": "boolean"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}